Cube Vertices

+++ Nade/src/Components/GameObject.h
@@ -25,13 +25,13 @@ namespace Nade {
 		void SetTexture(unsigned int texture) {
 			mData.texture = texture;
 		}
-		void Translate(float x, float y, float z) {
-			transform = glm::translate(transform, glm::vec3(x, y, z));
+		void Translate(double x, double y, double z) {
+			transform = glm::translate(transform, glm::vec3((float)x, (float)y, (float)z));
 		}
-		void Rotate(float x, float y, float z) {
-			transform = glm::rotate(transform, glm::radians(x), glm::vec3(1, 0, 0));
-			transform = glm::rotate(transform, glm::radians(y), glm::vec3(0, 1, 0));
-			transform = glm::rotate(transform, glm::radians(z), glm::vec3(0, 0, 1));
+		void Rotate(double x, double y, double z) {
+			transform = glm::rotate(transform, glm::radians((float)x), glm::vec3(1, 0, 0));
+			transform = glm::rotate(transform, glm::radians((float)y), glm::vec3(0, 1, 0));
+			transform = glm::rotate(transform, glm::radians((float)z), glm::vec3(0, 0, 1));
 		}
 		glm::mat4 transform = glm::mat4(1.0f);
 	private:
